LayoutTests/http/tests/websocket/tests/hybi/close-and-server-script-exception-expected.txt

Test how closing a connection triggers an exception in the python web socket server script or not

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws on master document is ready.
PASS insert a iframe, where open ws called 'socket1'
PASS 'socket1' is sent to the server. unload the iframe. expect receiving 'socket1' on ws...
PASS closedSocket is "Closed without exception"
PASS successfullyParsed is true

TEST COMPLETE
